Add files via upload
This program helps to solve easy Nonogram puzzles. Each row and column has only one clue. The program finds all possible ways to place the clue in the row or column. Then, it checks which cells are always filled in all options. These cells are marked as filled (1) in the result grid. The goal is to find the certain filled cells using only the given clues.